pokemon detail screen for ios

+++ pokedex/kmm/final/androidApp/src/main/java/dk/cngroup/pokedex/android/ui/navigation/graph/RootNavigationGraph.kt
@@ -42,7 +42,7 @@ fun RootNavigationGraph(
         ) {
             it.arguments?.getString(RootNavigationScreen.Parameter.POKEMON_ID)?.let { pokemonId ->
                 PokemonDetailScreen(
-                    viewModel = getViewModel { parametersOf(pokemonId.toInt()) }
+                    viewModel = getViewModel { parametersOf(pokemonId) }
                 )
             }
         }
